teacher

+++ Controllers/LessonsController.cs
@@ -9,7 +9,7 @@ using Microsoft.EntityFrameworkCore;
 namespace DistanceLearningSystem.Controllers;
 
 [ApiController]
-[Route("api/sections/{sectionId}/lessons")]
+[Route("api/lessons")]
 public class LessonsController : ControllerBase
 {
     private readonly ApplicationContext _context;
@@ -22,11 +22,20 @@ public class LessonsController : ControllerBase
     }
 
     [HttpGet("")]
-    public async Task<IActionResult> GetSectionLessons(int sectionId) =>
-        Ok(await _context.Lessons.Include(ls => ls.Section)
-            .Where(ls => ls.Section.SectionId == sectionId).ToListAsync());
+    public async Task<IActionResult> GetAllLessons()
+    {
+        var lessons = await _context.Lessons.ToListAsync();
+        return Ok(lessons);
+    }
 
-    [HttpPost("")]
+    [HttpGet("{lessonId}")]
+    public async Task<IActionResult> GetLessonById(int lessonId)
+    {
+        var lesson = await _context.Lessons.FindAsync(lessonId);
+        return Ok(lesson);
+    }
+
+    [HttpPost("sections/{sectionId}")]
     public async Task<IActionResult> AddLessonToSection(Lesson lesson, int sectionId)
     {
         var section = await _context.Sections.Include(sc => sc.Lessons)
@@ -35,31 +44,40 @@ public class LessonsController : ControllerBase
         section.Lessons.Add(lesson);
         _context.Update(section);
         await _context.SaveChangesAsync();
-        
+
         return Ok(section);
     }
-    
+
     [HttpPut("")]
     public async Task<IActionResult> EditLesson(Lesson lesson)
     {
         _context.Update(lesson);
         await _context.SaveChangesAsync();
-        
+
         return Ok();
     }
-    
-    [HttpDelete("")]
-    public async Task<IActionResult> DeleteLesson(Lesson lesson)
-    {
-        _context.Lessons.Remove(lesson);
-        await _context.SaveChangesAsync();
-        return Ok();
-    }
-    
-    [HttpGet("{lessonId}")]
-    public async Task<IActionResult> GetCourseSection(int lessonId)
+
+    [HttpDelete("{lessonId}")]
+    public async Task<IActionResult> DeleteLesson(int lessonId)
     {
         var lesson = await _context.Lessons.FindAsync(lessonId);
-        return Ok(lesson);
+
+        if (lesson == null)
+        {
+            return NotFound();
+        }
+
+        _context.Lessons.Remove(lesson);
+        await _context.SaveChangesAsync();
+
+        return Ok();
+    }
+    
+    [HttpGet("sections/{sectionId}")]
+    public async Task<IActionResult> GetSectionLessons(int sectionId)
+    {
+        var lessons = await _context.Lessons.Include(ls => ls.Section)
+            .Where(ls => ls.Section.SectionId == sectionId).ToListAsync();
+        return Ok(lessons);
     }
 }
